EFI: Final major version.
I need to thank to all who used and currently uses that EFI, it's been 4 whole years since I published that EFI to GitHub, at the beginning it was only for MSI motherboard, but as you may know: I don't have it anymore, so my focus went into Gigabyte motherboard. As you see in the commit title, that's the final major release of that EFI, since Tahoe is the last version for Intel CPUs. But don't worry, i'll keep it updated for every sub-release of Tahoe until the definitive end. Farewell, Hackintosh Enjoyers! It's been a good ride for me, and you all.

Signed-off-by: ArixElo <patroxgamer@gmail.com>